refactor (mesi_search): a module (candig.py) was wrongly situated outside
`candig.py` was outside the `mesi_search` package which was wrong because
we need the utility to be a part of that package.

Ideally, modules like `candig.py` would be a pluggable middleware instead
of having to write our own but that optimization is NOT happening here.
